% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardReexports.R
\name{bootplot}
\alias{bootplot}
\title{Produce delta objective function for boostrap}
\usage{
bootplot(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{Additional arguments passed to \code{\link[nlmixr2extra:bootplot]{nlmixr2extra::bootplot()}}.}
}
\value{
Fit traceplot or nothing.
}
\description{
Produce delta objective function for boostrap
}
\references{
R Niebecker,  MO Karlsson. (2013)
\emph{Are datasets for NLME models large enough for a bootstrap to provide reliable parameter uncertainty distributions?}
PAGE 2013.
\url{https://www.page-meeting.org/?abstract=2899}
}
\author{
Vipul Mann,  Matthew L. Fidler
}
