% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{.nlmixr2attach}
\alias{.nlmixr2attach}
\title{Load nlmixr2 stack}
\usage{
.nlmixr2attach(exclude = character())
}
\arguments{
\item{exclude}{exclude certain packages when loading.  This allows
package like babelmixr2 to not recursively load itself}
}
\value{
nothing, called for side effects
}
\description{
Load nlmixr2 stack
}
\author{
Matthew L. Fidler
}
\keyword{internal}
