% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{geom_cens}
\alias{geom_cens}
\title{Objects exported from other packages}
\value{
This returns a ggplot2 stat
}
\description{
This is a censoring geom that shows the left or right censoring specified in the nlmixr input data-set or fit
}
\details{
Requires the following aesthetics:
\itemize{
\item \code{x} Represents the independent variable, often the time scale
\item \code{y} represents the dependent variable
\item \code{CENS} for the censoring information; (\code{-1} right censored, \code{0} no censoring or \code{1} left censoring)
\item \code{LIMIT} which represents the corresponding limit ()
}

Will add boxes representing the areas of the fit that were censored.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2:stat_cens]{geom_cens}}}
}}

