% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{logit}
\alias{logit}
\title{Objects exported from other packages}
\value{
values from logit and expit
}
\description{
logit and inverse logit (expit) functions
}
\details{
logit is given by:

logit(p) = -log(1/p-1)

where:

p = x-low/high-low

expit is given by:

expit(p, low, high) = (high-low)/(1+exp(-alpha)) + low

The \code{logitNormInfo()} gives the mean, variance and coefficient of
variability on the untransformed scale.
}
\examples{
logit(0.25)

expit(-1.09)

logitNormInfo(logit(0.25), sd = 0.1)

logitNormInfo(logit(1, 0, 10), sd = 1, low = 0, high = 10)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2]{logit}}}
}}

