% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{model}
\alias{model}
\title{Model block for rxode2/nlmixr models (from \code{rxode2})
#' @inherit rxode2::model}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2]{model}}}
}}

