% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliasReexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nlmixr2NlmeControl}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nlmixr2est}{\code{\link[nlmixr2est]{nlmixr2NlmeControl}}}
}}

