% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{nlmixr2deps}
\alias{nlmixr2deps}
\title{List all nlmixr2 dependencies}
\usage{
nlmixr2deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
nlmixr2 packages.}
}
\description{
List all nlmixr2 dependencies
}
