% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{nlmixr2packages}
\alias{nlmixr2packages}
\title{List all packages in the nlmixr2}
\usage{
nlmixr2packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include nlmixr2 in the list?}
}
\description{
List all packages in the nlmixr2
}
\examples{
nlmixr2packages()
}
