% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{nlmixr2update}
\alias{nlmixr2update}
\title{Update nlmixr2 packages}
\usage{
nlmixr2update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
nlmixr2 packages.}
}
\description{
This will check to see if all nlmixr2 packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
nlmixr2update()
}
}
