% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardReexports.R
\name{profileFixedSingle}
\alias{profileFixedSingle}
\title{Estimate the objective function values for a model while fixing defined
parameter values}
\usage{
profileFixedSingle(fitted, which)
}
\arguments{
\item{fitted}{The fit model}

\item{which}{A data.frame with column names of parameters to fix and values
of the fitted value to fix (one row only).}
}
\value{
\code{which} with a column named \code{OFV} added with the objective function
value of the fitted estimate fixing the parameters in the other columns
}
\description{
Estimate the objective function values for a model while fixing defined
parameter values
}
\section{Functions}{

\itemize{
\item \code{profileFixedSingle()}: Estimate the objective function value for a model
while fixing a single set of defined parameter values (for use in parameter
profiling)

}
}

\seealso{
Other Profiling: 
\code{\link[nlmixr2extra]{fixedControl}()},
\code{\link[nlmixr2extra]{llpControl}()},
\code{\link[nlmixr2extra]{profile.nlmixr2FitCore}()},
\code{\link[nlmixr2extra]{profileLlp}()},
\code{\link[nlmixr2extra]{profileNlmixr2FitCoreRet}()}

Other Profiling: 
\code{\link[nlmixr2extra]{fixedControl}()},
\code{\link[nlmixr2extra]{llpControl}()},
\code{\link[nlmixr2extra]{profile.nlmixr2FitCore}()},
\code{\link[nlmixr2extra]{profileLlp}()},
\code{\link[nlmixr2extra]{profileNlmixr2FitCoreRet}()}
}
