% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardReexports.R
\name{profileLlp}
\alias{profileLlp}
\title{Profile confidence intervals with log-likelihood profiling}
\usage{
profileLlp(fitted, which, control)
}
\arguments{
\item{fitted}{The fit model}

\item{which}{Either \code{NULL} to profile all parameters or a character vector of
parameters to estimate}

\item{control}{A list passed to \code{llpControl()}}
}
\value{
A data.frame with columns named "Parameter" (the parameter name(s)
that were fixed), OFV (the objective function value), and the current
estimate for each of the parameters.  In addition, if any boundary is
found, the OFV increase will be indicated by the absolute value of the
"profileBound" column and if that boundary is the upper or lower boundary
will be indicated by the "profileBound" column being positive or negative,
respectively.
}
\description{
Profile confidence intervals with log-likelihood profiling
}
\seealso{
Other Profiling: 
\code{\link[nlmixr2extra]{fixedControl}()},
\code{\link[nlmixr2extra]{llpControl}()},
\code{\link[nlmixr2extra]{profile.nlmixr2FitCore}()},
\code{\link[nlmixr2extra]{profileFixed}()},
\code{\link[nlmixr2extra]{profileNlmixr2FitCoreRet}()}
}
