% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{lotri}
\alias{pdDiag}
\alias{pdSymm}
\alias{pdLogChol}
\alias{pdIdent}
\alias{pdCompSymm}
\alias{pdBlocked}
\alias{pdNatural}
\alias{pdConstruct}
\alias{pdFactor}
\alias{pdMat}
\alias{pdMatrix}
\alias{reStruct}
\alias{varWeights}
\alias{varPower}
\alias{varFixed}
\alias{varFunc}
\alias{varExp}
\alias{varConstPower}
\alias{varIdent}
\alias{varComb}
\alias{groupedData}
\alias{getData}
\alias{nlme}
\alias{ACF}
\alias{VarCorr}
\alias{getVarCov}
\alias{augPred}
\alias{fixef}
\alias{fixed.effects}
\alias{ranef}
\alias{random.effects}
\alias{.nlmixrNlmeFun}
\title{Objects exported from other packages}
\value{
A list of table options for nlmixr2
}
\description{
Output table/data.frame options
}
\details{
If you ever want to add CWRES/FOCEi objective function you can use the \code{\link[nlmixr2est]{addCwres}}

If you ever want to add NPDE/EPRED columns you can use the \code{\link[nlmixr2est]{addNpde}}
}
\author{
Matthew L. Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{nlmixr2est}{\code{\link[nlmixr2est:dot-nlmixrNlmeFun]{.nlmixrNlmeFun}}, \code{\link[nlmixr2est:reexports]{ACF}}, \code{\link[nlmixr2est:reexports]{VarCorr}}, \code{\link[nlmixr2est:reexports]{augPred}}, \code{\link[nlmixr2est:reexports]{fixed.effects}}, \code{\link[nlmixr2est:reexports]{fixef}}, \code{\link[nlmixr2est:reexports]{getData}}, \code{\link[nlmixr2est:reexports]{getVarCov}}, \code{\link[nlmixr2est:reexports]{groupedData}}, \code{\link[nlmixr2est:reexports]{nlme}}, \code{\link[nlmixr2est:reexports]{pdBlocked}}, \code{\link[nlmixr2est:reexports]{pdCompSymm}}, \code{\link[nlmixr2est:reexports]{pdConstruct}}, \code{\link[nlmixr2est:reexports]{pdDiag}}, \code{\link[nlmixr2est:reexports]{pdFactor}}, \code{\link[nlmixr2est:reexports]{pdIdent}}, \code{\link[nlmixr2est:reexports]{pdLogChol}}, \code{\link[nlmixr2est:reexports]{pdMat}}, \code{\link[nlmixr2est:reexports]{pdMatrix}}, \code{\link[nlmixr2est:reexports]{pdNatural}}, \code{\link[nlmixr2est:reexports]{pdSymm}}, \code{\link[nlmixr2est:reexports]{random.effects}}, \code{\link[nlmixr2est:reexports]{ranef}}, \code{\link[nlmixr2est:reexports]{reStruct}}, \code{\link[nlmixr2est:reexports]{varComb}}, \code{\link[nlmixr2est:reexports]{varConstPower}}, \code{\link[nlmixr2est:reexports]{varExp}}, \code{\link[nlmixr2est:reexports]{varFixed}}, \code{\link[nlmixr2est:reexports]{varFunc}}, \code{\link[nlmixr2est:reexports]{varIdent}}, \code{\link[nlmixr2est:reexports]{varPower}}, \code{\link[nlmixr2est:reexports]{varWeights}}}

  \item{rxode2}{\code{\link[rxode2:reexports]{lotri}}}
}}

