% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{rxCat}
\alias{rxCat}
\title{Objects exported from other packages}
\value{
nothing
}
\description{
Use cat when rxode2.verbose is TRUE
}
\author{
Matthew L. Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2]{rxCat}}}
}}

