% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{rxParams}
\alias{rxParams}
\alias{rxParam}
\title{Parameters specified by the model (from \code{rxode2})}
\value{
When extracting the parameters from an rxode2 model, a
character vector listing the parameters in the model.
}
\description{
This returns the model's parameters that are required to solve the
ODE system, and can be used to pipe parameters into an rxode2 solve
}
\seealso{
Other Query model information: 
\code{\link[rxode2]{rxDfdy}()},
\code{\link[rxode2]{rxInits}()},
\code{\link[rxode2]{rxLhs}()},
\code{\link[rxode2]{rxModelVars}()},
\code{\link[rxode2]{rxState}()}
}
\author{
Matthew L.Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2:rxParams]{rxParam}}, \code{\link[rxode2]{rxParams}}}
}}

