% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{rxSolve}
\alias{rxSolve}
\title{Objects exported from other packages}
\value{
An \dQuote{rxSolve} solve object that stores the solved
value in a special data.frame or other type as determined by
\code{returnType}. By default this has as many rows as there are
sampled time points and as many columns as system variables (as
defined by the ODEs and additional assignments in the rxode2 model
code).  It also stores information about the call to allow
dynamic updating of the solved object.

The operations for the object are similar to a data-frame, but
expand the \code{$} and \verb{[[""]]} access operators and assignment
operators to resolve based on different parameter values, initial
conditions, solver parameters, or events (by updating the \code{time}
variable).

You can call the \code{\link[rxode2:eventTable]{eventTable()}} methods on the solved object to
update the event table and resolve the system of equations.
}
\description{
This uses rxode2 family of objects, file, or model specification to
solve a ODE system.  There are many options for a solved rxode2
model, the first are the required \code{object}, and \code{events} with the
some-times optional \code{params} and \code{inits}.
}
\details{
The rest of the document focus on the different ODE solving
methods, followed by the core solving method's options, rxode2 event
handling options, rxode2's numerical stability options, rxode2's
output options, and finally internal rxode2 options or compatibility
options.
}
\references{
"New Scaling and Squaring Algorithm for the Matrix Exponential", by
Awad H. Al-Mohy and Nicholas J. Higham, August 2009

Roger B. Sidje (1998).  EXPOKIT: Software package for computing
matrix exponentials.  ACM - Transactions on Mathematical Software
\emph{24}(1), 130-156.

Hindmarsh, A. C.
\emph{ODEPACK, A Systematized Collection of ODE Solvers}.
Scientific Computing, R. S. Stepleman et al. (Eds.),
North-Holland, Amsterdam, 1983, pp. 55-64.

Petzold, L. R.
\emph{Automatic Selection of Methods for Solving Stiff and Nonstiff
Systems of Ordinary Differential Equations}.
Siam J. Sci. Stat. Comput. 4 (1983), pp. 136-148.

Hairer, E., Norsett, S. P., and Wanner, G.
\emph{Solving ordinary differential equations I, nonstiff problems}.
2nd edition, Springer Series in Computational Mathematics,
Springer-Verlag (1993).
}
\seealso{
\code{\link[rxode2:rxode2]{rxode2()}}
}
\author{
Matthew Fidler, Melissa Hallow and  Wenping Wang
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2]{rxSolve}}}
}}

