% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{saemControl}
\alias{saemControl}
\title{Control options for the \code{saem} procedure (from \code{nlmixr2est})}
\value{
List of options to be used in \code{\link[nlmixr2est]{nlmixr2}} fit for
    SAEM.
}
\description{
Control Options for SAEM
}
\author{
Wenping Wang & Matthew L. Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nlmixr2est}{\code{\link[nlmixr2est]{saemControl}}}
}}

