% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{setOfv}
\alias{setOfv}
\title{Objects exported from other packages}
\value{
Nothing
}
\description{
Set/get Objective function type for a nlmixr2 object
}
\author{
Matthew L. Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nlmixr2est}{\code{\link[nlmixr2est]{setOfv}}}
}}

