% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{stat_amt}
\alias{stat_amt}
\title{Objects exported from other packages}
\value{
This returns a stat_amt in context of a ggplot2 plot
}
\description{
This is a dosing geom that shows the vertical lines where a dose occurs
}
\details{
Requires the following aesthetics:

\itemize{
\item x representing the x values, usually time
\item amt representing the dosing values;  They are missing or zero when no dose is given
}
}
\examples{
library(rxode2)
library(units)


## Model from RxODE tutorial
mod1 <-rxode2({
       KA=2.94E-01
       CL=1.86E+01
       V2=4.02E+01
       Q=1.05E+01
       V3=2.97E+02
       Kin=1
       Kout=1
       EC50=200
       C2 = centr/V2
       C3 = peri/V3
       d/dt(depot) =-KA*depot
       d/dt(centr) = KA*depot - CL*C2 - Q*C2 + Q*C3
       d/dt(peri)  =                    Q*C2 - Q*C3
       d/dt(eff)  = Kin - Kout*(1-C2/(EC50+C2))*eff
})

## These are making the more complex regimens of the rxode2 tutorial

## bid for 5 days
bid <- et(timeUnits="hr") \%>\%
      et(amt=10000,ii=12,until=set_units(5, "days"))

## qd for 5 days
qd <- et(timeUnits="hr") \%>\%
      et(amt=20000,ii=24,until=set_units(5, "days"))

## bid for 5 days followed by qd for 5 days

et <- seq(bid,qd) \%>\% et(seq(0,11*24,length.out=100));

bidQd <- rxSolve(mod1, et, addDosing=TRUE)

# by default dotted and under-stated
plot(bidQd, C2) + geom_amt(aes(amt=amt))

# of course you can make it a bit more visible

plot(bidQd, C2) + geom_amt(aes(amt=amt), col="red", lty=1, size=1.2)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2]{stat_amt}}}
}}

