% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{model}
\alias{model}
\title{Objects exported from other packages}
\value{
Model block with ini information included.  \code{ini} must be
called before \code{model} block
}
\description{
Model block for rxode2/nlmixr models
}
\author{
Matthew Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rxode2}{\code{\link[rxode2]{model}}}
}}

