% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{nlmixr2NlmeControl}
\alias{nlmixr2NlmeControl}
\title{Control options for the \code{nlme} procedure (from \code{nlmixr2est})}
\description{
The values supplied in the function call replace the defaults and
a list with all possible arguments is returned.  The returned list
is used as the ‘control’ argument to the ‘nlme’ function.
}
\examples{
nlmixr2est::nlmeControl()
nlmixr2NlmeControl()
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nlmixr2est}{\code{\link[nlmixr2est]{nlmixr2NlmeControl}}}
}}

