% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{rxode2}
\alias{rxode}
\alias{RxODE}
\alias{ini}
\alias{model}
\alias{lotri}
\alias{logit}
\alias{expit}
\alias{probit}
\alias{probitInv}
\alias{rxSolve}
\alias{rxClean}
\alias{rxCat}
\alias{eventTable}
\alias{add.dosing}
\alias{add.sampling}
\alias{pdDiag}
\alias{pdSymm}
\alias{pdLogChol}
\alias{pdIdent}
\alias{pdCompSymm}
\alias{pdBlocked}
\alias{pdNatural}
\alias{pdConstruct}
\alias{pdFactor}
\alias{pdMat}
\alias{pdMatrix}
\alias{reStruct}
\alias{varWeights}
\alias{varPower}
\alias{varFixed}
\alias{varFunc}
\alias{varExp}
\alias{varConstPower}
\alias{varIdent}
\alias{varComb}
\alias{groupedData}
\alias{getData}
\alias{et}
\alias{rxParams}
\alias{rxParam}
\alias{geom_cens}
\alias{geom_amt}
\alias{stat_cens}
\alias{stat_amt}
\alias{rxControl}
\alias{traceplot}
\alias{vpcPlot}
\alias{nlmixr2}
\alias{nlmixr}
\alias{saemControl}
\alias{foceiControl}
\alias{nlmixr2NlmeControl}
\alias{nlmeControl}
\alias{tableControl}
\alias{nlme}
\alias{ACF}
\alias{VarCorr}
\alias{getVarCov}
\alias{augPred}
\alias{fixef}
\alias{fixed.effects}
\alias{ranef}
\alias{random.effects}
\alias{.nlmixrNlmeFun}
\alias{addCwres}
\alias{addNpde}
\alias{addTable}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{nlmixr2est}{\code{\link[nlmixr2est:dot-nlmixrNlmeFun]{.nlmixrNlmeFun}}, \code{\link[nlmixr2est:reexports]{ACF}}, \code{\link[nlmixr2est:reexports]{VarCorr}}, \code{\link[nlmixr2est]{addCwres}}, \code{\link[nlmixr2est]{addNpde}}, \code{\link[nlmixr2est]{addNpde}}, \code{\link[nlmixr2est]{addTable}}, \code{\link[nlmixr2est:reexports]{augPred}}, \code{\link[nlmixr2est:reexports]{fixed.effects}}, \code{\link[nlmixr2est:reexports]{fixef}}, \code{\link[nlmixr2est]{foceiControl}}, \code{\link[nlmixr2est:reexports]{getData}}, \code{\link[nlmixr2est:reexports]{getVarCov}}, \code{\link[nlmixr2est:reexports]{groupedData}}, \code{\link[nlmixr2est:reexports]{nlme}}, \code{\link[nlmixr2est:nlmixr2NlmeControl]{nlmeControl}}, \code{\link[nlmixr2est:nlmixr2]{nlmixr}}, \code{\link[nlmixr2est]{nlmixr2}}, \code{\link[nlmixr2est]{nlmixr2NlmeControl}}, \code{\link[nlmixr2est:reexports]{pdBlocked}}, \code{\link[nlmixr2est:reexports]{pdCompSymm}}, \code{\link[nlmixr2est:reexports]{pdConstruct}}, \code{\link[nlmixr2est:reexports]{pdDiag}}, \code{\link[nlmixr2est:reexports]{pdFactor}}, \code{\link[nlmixr2est:reexports]{pdIdent}}, \code{\link[nlmixr2est:reexports]{pdLogChol}}, \code{\link[nlmixr2est:reexports]{pdMat}}, \code{\link[nlmixr2est:reexports]{pdMatrix}}, \code{\link[nlmixr2est:reexports]{pdNatural}}, \code{\link[nlmixr2est:reexports]{pdSymm}}, \code{\link[nlmixr2est:reexports]{random.effects}}, \code{\link[nlmixr2est:reexports]{ranef}}, \code{\link[nlmixr2est:reexports]{reStruct}}, \code{\link[nlmixr2est]{saemControl}}, \code{\link[nlmixr2est]{tableControl}}, \code{\link[nlmixr2est:reexports]{varComb}}, \code{\link[nlmixr2est:reexports]{varConstPower}}, \code{\link[nlmixr2est:reexports]{varExp}}, \code{\link[nlmixr2est:reexports]{varFixed}}, \code{\link[nlmixr2est:reexports]{varFunc}}, \code{\link[nlmixr2est:reexports]{varIdent}}, \code{\link[nlmixr2est:reexports]{varPower}}, \code{\link[nlmixr2est:reexports]{varWeights}}}

  \item{nlmixr2plot}{\code{\link[nlmixr2plot:reexports]{geom_amt}}, \code{\link[nlmixr2plot:reexports]{geom_cens}}, \code{\link[nlmixr2plot:reexports]{stat_amt}}, \code{\link[nlmixr2plot:reexports]{stat_cens}}, \code{\link[nlmixr2plot]{traceplot}}, \code{\link[nlmixr2plot]{traceplot}}, \code{\link[nlmixr2plot]{vpcPlot}}}

  \item{rxode2}{\code{\link[rxode2:rxode2]{RxODE}}, \code{\link[rxode2]{add.dosing}}, \code{\link[rxode2]{add.sampling}}, \code{\link[rxode2]{et}}, \code{\link[rxode2]{et}}, \code{\link[rxode2]{eventTable}}, \code{\link[rxode2:logit]{expit}}, \code{\link[rxode2:stat_amt]{geom_amt}}, \code{\link[rxode2:stat_cens]{geom_cens}}, \code{\link[rxode2]{ini}}, \code{\link[rxode2]{logit}}, \code{\link[rxode2:reexports]{lotri}}, \code{\link[rxode2]{model}}, \code{\link[rxode2]{probit}}, \code{\link[rxode2:probit]{probitInv}}, \code{\link[rxode2]{rxCat}}, \code{\link[rxode2]{rxClean}}, \code{\link[rxode2:rxSolve]{rxControl}}, \code{\link[rxode2:rxParams]{rxParam}}, \code{\link[rxode2]{rxParams}}, \code{\link[rxode2]{rxSolve}}, \code{\link[rxode2:rxode2]{rxode}}, \code{\link[rxode2]{rxode2}}, \code{\link[rxode2]{stat_amt}}, \code{\link[rxode2]{stat_cens}}}
}}

