% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardReexports.R
\name{setOfv}
\alias{setOfv}
\title{Set/get Objective function type for a nlmixr2 object}
\usage{
setOfv(x, type)
}
\arguments{
\item{x}{nlmixr2 fit object}

\item{type}{Type of objective function to use for AIC, BIC, and
$objective}
}
\value{
Nothing
}
\description{
Set/get Objective function type for a nlmixr2 object
}
\author{
Matthew L. Fidler
}
