% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{tableControl}
\alias{tableControl}
\title{Table output control options for the \code{nlme} procedure (from \code{nlmixr2est})}
\value{
A list of table options for nlmixr2
}
\description{
Output table/data.frame options
}
\details{
If you ever want to add CWRES/FOCEi objective function you can use the \code{\link[nlmixr2est]{addCwres}}

If you ever want to add NPDE/EPRED columns you can use the \code{\link[nlmixr2est]{addNpde}}
}
\author{
Matthew L. Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nlmixr2est}{\code{\link[nlmixr2est]{tableControl}}}
}}

