% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{traceplot}
\alias{traceplot}
\title{Produce trace-plot for fit if applicable (from \code{nlmixr2plot})}
\value{
Fit traceplot or nothing.
}
\description{
Produce trace-plot for fit if applicable
}
\author{
Rik Schoemaker, Wenping Wang & Matthew L. Fidler
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{nlmixr2plot}{\code{\link[nlmixr2plot]{traceplot}}}
}}

