% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardReexports.R
\name{traceplot}
\alias{traceplot}
\title{Produce trace-plot for fit if applicable}
\usage{
traceplot(x, ...)
}
\arguments{
\item{x}{fit object}

\item{...}{Additional arguments passed to \code{\link[nlmixr2plot:traceplot]{nlmixr2plot::traceplot()}}.}
}
\value{
Fit traceplot or nothing.
}
\description{
Produce trace-plot for fit if applicable
}
\author{
Rik Schoemaker, Wenping Wang & Matthew L. Fidler
}
